refactor: simplify NotesPage by integrating NotesEditor component and optimizing state management with useCallback
